refactor: organize checklist components
